MultiplicativeGroupAlgebra.CreateCryptoGroup skips more known non-prime candidates.
Sophie Germain primes can only have a value of 5 modulo 6, which is used to improve the prime finding algorithm in MultiplicativeGroupAlgebra.CreateCryptoGroup, which previously was checking every odd number.
This reduces the number of primality tests that are performed by up to 2/3.

Also joins the previously separate test cases into a single implementation and fixes a bug in which the rng mock was
not returning the expected bytes.